refactor: property를 ConfigurationProperty 방식으로 개선
- 깨지는 test 해결
