+++ src/test/java/com/nexters/phochak/conf/EmbeddedRedisConfig.java
@@ -1,6 +1,5 @@
-package com.nexters.phochak.config;
+package com.nexters.phochak.conf;
 
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
@@ -13,7 +12,6 @@ import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 
-@Slf4j
 @Profile("test")
 @Configuration
 public class EmbeddedRedisConfig {
